+++ projects/minecraft/v1.19.2/vanilla/src/main/java/kittify/vanilla/common/integration/trinkets/CuteCollarTrinket.java
@@ -1,0 +1,27 @@
+package kittify.vanilla.common.integration.trinkets;
+
+import com.google.common.collect.Multimap;
+import dev.emi.trinkets.api.SlotReference;
+import dev.emi.trinkets.api.Trinket;
+import dev.emi.trinkets.api.TrinketEnums;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attribute;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.item.ItemStack;
+
+import java.util.UUID;
+
+public class CuteCollarTrinket implements Trinket {
+    @Override
+    public Multimap<Attribute, AttributeModifier> getModifiers(ItemStack stack, SlotReference slot, LivingEntity entity, UUID uuid) {
+        Multimap<Attribute, AttributeModifier> modifiers = Trinket.super.getModifiers(stack, slot, entity, uuid);
+        modifiers.put(Attributes.MOVEMENT_SPEED, new AttributeModifier(uuid, "kittify:cute_collar_speed", 2, AttributeModifier.Operation.MULTIPLY_TOTAL));
+        return modifiers;
+    }
+
+    @Override
+    public TrinketEnums.DropRule getDropRule(ItemStack stack, SlotReference slot, LivingEntity entity) {
+        return TrinketEnums.DropRule.KEEP;
+    }
+}