+++ projects/minecraft/v1.19.2/vanilla/src/main/java/kittify/vanilla/common/integration/trinkets/KittifyTrinketsIntegration.java
@@ -1,0 +1,13 @@
+package kittify.vanilla.common.integration.trinkets;
+
+import dev.emi.trinkets.api.TrinketsApi;
+import kittify.vanilla.common.registry.KittifyItems;
+
+public class KittifyTrinketsIntegration {
+    private KittifyTrinketsIntegration() {
+    }
+
+    public static void afterRegisterItems() {
+        TrinketsApi.registerTrinket(KittifyItems.CUTE_COLLAR, new CuteCollarTrinket());
+    }
+}